// src/app/models/epg-layout.ts
export class EpgLayout {
  days_date_format: string;
  id: number;
  name: string;
  navigation_type: string;
  number_channels: number;
  number_days_backward: number;
  number_days_forward: number;
  number_hours: number;
  program_selection_behaviour: string;
  show_channel_number: number;
  show_genres: number;
  show_preview: string;
  show_selected_program_info: string;
  slug: string;
  time_slot_duration: number;
  type: string;
}